Adding Miller-Rabin prime checker.  Needs more testing, but out of time.  Also adding a Compiled copy to source control.

+++ MaxPrimeNumberCalculator_Tests/ExponentialStepPrimeFinder_Tests.cs
@@ -15,6 +15,7 @@
 #endregion
 
 using System;
+using System.Numerics;
 using MaxPrimeNumberCalculator.Interfaces;
 using MaxPrimeNumberCalculator.PrimeNumberFinders;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
@@ -30,7 +31,7 @@ namespace MaxPrimeNumberCalculator_Tests
         {
             var mockPrimeChecker = new Mock<IPrimeChecker>();
             //Mock out the prime checker to always return true
-            mockPrimeChecker.Setup(mpc => mpc.IsPrime(It.IsAny<ulong>())).Returns(true);
+            mockPrimeChecker.Setup(mpc => mpc.IsPrime(It.IsAny<BigInteger>())).Returns(true);
             var primeFinder = new ExponentialStepPrimeFinder(5, 5, mockPrimeChecker.Object);
 
             var primeOne = primeFinder.GetNextLargestPrime();
